+++ plugin.xml
@@ -1,5 +1,5 @@
 <library path="src">
-  <class name="Gauges2" type="rqt_gauges_2.gauges_2.Gauges2" base_class_type="rqt_gui_py::Plugin">
+  <class name="Speedometer" type="rqt_gauges_2.gauges_2.Speedometer" base_class_type="rqt_gui_py::Plugin">
     <description>
       A Visualization tool to keep track of the values of sensors
     </description>
