+++ resource/speedometer.ui
@@ -6,17 +6,20 @@
    <rect>
     <x>0</x>
     <y>0</y>
-    <width>490</width>
+    <width>479</width>
     <height>509</height>
    </rect>
   </property>
   <property name="windowTitle">
    <string>Speedometer</string>
   </property>
+  <property name="autoFillBackground">
+   <bool>false</bool>
+  </property>
   <widget class="QLabel" name="min_value_label">
    <property name="geometry">
     <rect>
-     <x>50</x>
+     <x>40</x>
      <y>470</y>
      <width>81</width>
      <height>31</height>
@@ -34,27 +37,49 @@
   <widget class="QTextEdit" name="min_value">
    <property name="geometry">
     <rect>
-     <x>130</x>
+     <x>120</x>
      <y>470</y>
-     <width>81</width>
+     <width>51</width>
      <height>31</height>
     </rect>
+   </property>
+   <property name="placeholderText">
+    <string/>
    </property>
   </widget>
   <widget class="QTextEdit" name="max_value">
    <property name="geometry">
     <rect>
-     <x>360</x>
+     <x>270</x>
      <y>470</y>
-     <width>81</width>
+     <width>51</width>
      <height>31</height>
     </rect>
+   </property>
+   <property name="font">
+    <font>
+     <weight>50</weight>
+     <italic>false</italic>
+     <bold>false</bold>
+     <underline>false</underline>
+     <strikeout>false</strikeout>
+    </font>
+   </property>
+   <property name="html">
+    <string>&lt;!DOCTYPE HTML PUBLIC &quot;-//W3C//DTD HTML 4.0//EN&quot; &quot;http://www.w3.org/TR/REC-html40/strict.dtd&quot;&gt;
+&lt;html&gt;&lt;head&gt;&lt;meta name=&quot;qrichtext&quot; content=&quot;1&quot; /&gt;&lt;style type=&quot;text/css&quot;&gt;
+p, li { white-space: pre-wrap; }
+&lt;/style&gt;&lt;/head&gt;&lt;body style=&quot; font-family:'Ubuntu'; font-size:11pt; font-weight:400; font-style:normal;&quot;&gt;
+&lt;p style=&quot;-qt-paragraph-type:empty; margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;&quot;&gt;&lt;br /&gt;&lt;/p&gt;&lt;/body&gt;&lt;/html&gt;</string>
+   </property>
+   <property name="placeholderText">
+    <string/>
    </property>
   </widget>
   <widget class="QLabel" name="max_value_label">
    <property name="geometry">
     <rect>
-     <x>270</x>
+     <x>180</x>
      <y>476</y>
      <width>81</width>
      <height>21</height>
@@ -110,7 +135,7 @@
     </rect>
    </property>
   </widget>
-  <widget class="QWidget" name="speedometer_widget" native="true">
+  <widget class="SpeedometerGauge" name="speedometer_gauge" native="true">
    <property name="geometry">
     <rect>
      <x>0</x>
@@ -120,7 +145,62 @@
     </rect>
    </property>
   </widget>
+  <widget class="QLabel" name="units_label">
+   <property name="geometry">
+    <rect>
+     <x>330</x>
+     <y>476</y>
+     <width>81</width>
+     <height>21</height>
+    </rect>
+   </property>
+   <property name="font">
+    <font>
+     <pointsize>12</pointsize>
+    </font>
+   </property>
+   <property name="text">
+    <string>Units:</string>
+   </property>
+  </widget>
+  <widget class="QTextEdit" name="units">
+   <property name="geometry">
+    <rect>
+     <x>380</x>
+     <y>470</y>
+     <width>51</width>
+     <height>31</height>
+    </rect>
+   </property>
+   <property name="font">
+    <font>
+     <weight>50</weight>
+     <italic>false</italic>
+     <bold>false</bold>
+     <underline>false</underline>
+     <strikeout>false</strikeout>
+    </font>
+   </property>
+   <property name="html">
+    <string>&lt;!DOCTYPE HTML PUBLIC &quot;-//W3C//DTD HTML 4.0//EN&quot; &quot;http://www.w3.org/TR/REC-html40/strict.dtd&quot;&gt;
+&lt;html&gt;&lt;head&gt;&lt;meta name=&quot;qrichtext&quot; content=&quot;1&quot; /&gt;&lt;style type=&quot;text/css&quot;&gt;
+p, li { white-space: pre-wrap; }
+&lt;/style&gt;&lt;/head&gt;&lt;body style=&quot; font-family:'Ubuntu'; font-size:11pt; font-weight:400; font-style:normal;&quot;&gt;
+&lt;p style=&quot;-qt-paragraph-type:empty; margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;&quot;&gt;&lt;br /&gt;&lt;/p&gt;&lt;/body&gt;&lt;/html&gt;</string>
+   </property>
+   <property name="placeholderText">
+    <string/>
+   </property>
+  </widget>
  </widget>
+ <customwidgets>
+  <customwidget>
+   <class>SpeedometerGauge</class>
+   <extends>QWidget</extends>
+   <header>rqt_gauges_2.speedometer_gauge.h</header>
+   <container>1</container>
+  </customwidget>
+ </customwidgets>
  <resources/>
  <connections/>
 </ui>
