+++ src/rqt_gauges_2/main.py
@@ -1,11 +1,11 @@
 import sys
 
 from rqt_gui.main import Main
-from rqt_gauges_2.gauges_2 import Gauges2
+from rqt_gauges_2.speedometer import Speedometer
 
 
 def main():
-    plugin = 'rqt_gauges_2.gauges_2.Gauges2'
+    plugin = 'rqt_gauges_2.gauges_2.Speedometer'
     main = Main(filename=plugin)
     sys.exit(main.main(sys.argv, standalone=plugin))
